Enhance OTEL Collector support in the inventory API
- Updated the AddAgentOTELCollectorCommand to handle syslog port as int64.
- Improved API documentation to include OTEL Collector in agent types and descriptions.
- Added validation and configuration for OTEL Collector logs, including log sources.
- Enhanced JSON schemas and validation for OTEL Collector in various API responses.
- Introduced new structures for handling OTEL Collector parameters in requests and responses.

This update lays the foundation for comprehensive OTEL Collector integration, facilitating logs, traces, and future telemetry types.

+++ admin/commands/inventory/add_agent_otel_collector.go
@@ -92,7 +92,7 @@ func (cmd *AddAgentOTELCollectorCommand) RunCmd() (commands.Result, error) {
 			Enabled:        cmd.EnableLogs || len(logSources) > 0,
 			LogSources:     logSources,
 			EnableSyslog:   cmd.EnableSyslog,
-			SyslogPort:     cmd.SyslogPort,
+			SyslogPort:     int64(cmd.SyslogPort),
 			EnableJournald: cmd.EnableJournald,
 			JournaldUnits:  cmd.JournaldUnits,
 		}
@@ -117,4 +117,3 @@ func (cmd *AddAgentOTELCollectorCommand) RunCmd() (commands.Result, error) {
 		Agent: resp.Payload.OtelCollector,
 	}, nil
 }
-
